move redshift data implementation queryrunenr/redshiftdata package
